tensor: Fix printing of F64 elements

+++ le/letensor.c
@@ -693,7 +693,7 @@ le_tensor_to_cstr(LeTensor *self)
                     sprintf(ptr, "%1.3f%n", ((float *)self->data)[y * self->shape->sizes[1] + x], &written);
                     break;
                 case LE_TYPE_FLOAT64:
-                    sprintf(ptr, "%1.3lf%n", ((float *)self->data)[y * self->shape->sizes[1] + x], &written);
+                    sprintf(ptr, "%1.3lf%n", ((double *)self->data)[y * self->shape->sizes[1] + x], &written);
                     break;
                 case LE_TYPE_VOID:
                 default:
